pyproject.toml
[project]
name = "gausstr-lightning"
version = "0.1.0"
description = "PyTorch Lightning implementation of GaussTR for self-supervised 3D spatial understanding"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Haoyi Jiang"},
    {name = "Max-Bin"},
]
keywords = ["3d-vision", "occupancy-prediction", "gaussian-splatting", "pytorch-lightning"]

dependencies = [
    # Core PyTorch
    "torch>=2.0.0",
    "torchvision>=0.15.0",

    # PyTorch Lightning
    "pytorch-lightning>=2.0.0",
    "lightning>=2.0.0",

    # Configuration management
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",

    # Metrics
    "torchmetrics>=1.0.0",

    # Data processing
    "numpy>=1.20.0",
    "pillow>=9.0.0",
    "pyquaternion>=0.9.0",

    # Gaussian splatting
    "gsplat>=1.0.0",

    # Logging
    "mlflow>=2.0.0",

    # Progress bar and UI
    "rich>=12.0.0",
    "tqdm>=4.60.0",
    "terminaltables>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
tensorboard = [
    "tensorboard>=2.10.0",
]
wandb = [
    "wandb>=0.13.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
